Functions for dumping spec and graphql moved to testkit package
